# openapi.yaml
openapi: 3.0.3
info:
  title: Modullar Advancia Healthcare API
  description: Healthcare payment and compliance management platform
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /disputes:
    get:
      summary: List all disputes
      tags: [Disputes]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of disputes
    post:
      summary: Create dispute
      tags: [Disputes]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dispute_number, dispute_reason, amount]
              properties:
                dispute_number:
                  type: string
                transaction_id:
                  type: string
                  format: uuid
                dispute_reason:
                  type: string
                  enum: [fraud, duplicate, product_not_received, service_not_provided, other]
                amount:
                  type: number
                customer_description:
                  type: string
      responses:
        '201':
          description: Dispute created
...
  /stripe/customers:
    post:
      summary: Create Stripe customer
      tags: [Stripe Payments]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                name:
                  type: string
                metadata:
                  type: object
      responses:
        '200':
          description: Customer created
  /stripe/payment-intents:
    post:
      summary: Create payment intent
      tags: [Stripe Payments]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  minimum: 50
                currency:
                  type: string
                  default: usd
                customerId:
                  type: string
                description:
                  type: string
                metadata:
                  type: object
      responses:
        '200':
          description: Payment intent created
  /stripe/payment-intents/{id}:
    get:
      summary: Get payment intent
      tags: [Stripe Payments]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment intent details
  /stripe/checkout/sessions:
    post:
      summary: Create checkout session
      tags: [Stripe Payments]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lineItems]
              properties:
                lineItems:
                  type: array
                  items:
                    type: object
                customerId:
                  type: string
                successUrl:
                  type: string
                cancelUrl:
                  type: string
      responses:
        '200':
          description: Checkout session created
  /stripe/refunds:
    post:
      summary: Create refund
      tags: [Stripe Payments]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paymentIntentId]
              properties:
                paymentIntentId:
                  type: string
                amount:
                  type: number
                reason:
                  type: string
      responses:
        '200':
          description: Refund created

  # Stripe Connect Routes
...